+++ src/app/(platform)/layout.tsx
@@ -35,7 +35,7 @@ const PlatformLayout = async ({
       </div>
 
       <main className="relative w-full p-4 pt-16 md:p-0 md:pt-1">
-        <SidebarTrigger className="absolute hidden md:block md:left-6 md:top-6" />
+        <SidebarTrigger className="absolute hidden md:left-6 md:top-6 md:block" />
         {children}
       </main>
     </SidebarProvider>
