+++ src/app/(platform)/perfil/page.tsx
