+++ src/components/home/advises-card.tsx
@@ -5,5 +5,7 @@ import { StatCard } from './stat-card';
 export const AdvisesCard = async () => {
   const numberOfAdvises = await prisma.advise.count({});
 
-  return <StatCard href="/consejos" title="Consejos" Icon={MessageSquare} value={numberOfAdvises} />;
+  return (
+    <StatCard href="/consejos" title="Consejos" Icon={MessageSquare} value={numberOfAdvises} />
+  );
 };
