+++ src/middleware.ts
@@ -1,16 +1,16 @@
-import { NextResponse } from 'next/server';
 import type { NextRequest } from 'next/server';
+import { NextResponse } from 'next/server';
 
 export function middleware(request: NextRequest) {
   const sessionId = request.cookies.get('sessionId')?.value;
 
   if (!sessionId) {
-    return NextResponse.redirect(new URL('/auth/sign-in', request.url));
+    return NextResponse.redirect(new URL('/autenticacion/iniciar-sesion', request.url));
   }
 
   return NextResponse.next();
 }
 
 export const config = {
-  matcher: ['/profile/:path*'],
+  matcher: ['/perfil/:path*'],
 };
